Don't checkout if BRANCH

+++ kazoo/build/build.sh
@@ -2,9 +2,14 @@
 [ -n "$SKIP_BUILD" ] && exit 0
 . erlang/activate
 cd kazoo
-git fetch
-COMMIT=$(cat ~/commit)
-git reset --hard $COMMIT
-git clean -d -f
+
+if [ -z $BRANCH ]
+then
+	git fetch
+	COMMIT=$(cat ~/commit)
+	git reset --hard $COMMIT
+	git clean -d -f
+fi
+
 cp ../sys.config rel/
 make compile build-dev-release
